drop old_timestamps_manager
